% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.ridge.lm.R
\name{dual.ridge.lm}
\alias{dual.ridge.lm}
\title{Fitting Linear Models with Ridge Penalty using Dual Solution}
\usage{
dual.ridge.lm(
  formula,
  data,
  lambda,
  intercept = F,
  standardize = F,
  kernel = "lin",
  const = 0,
  degree = NULL,
  scale = NULL
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}: a symbolic
description of the model to be fitted following the standard of
\code{\link[stats]{lm}}.}

\item{data}{an optional data frame, list or environment (or object coercible
by \code{\link[base]{as.data.frame}} to a data frame) containing the
variables in the model. If not found in data, the variables are taken from
environment (\code{formula}), typically the environment from which this
function is called.}

\item{lambda}{penalty term scaling hyperparameter.}

\item{intercept}{optional boolean indicating whether to fit an intercept. If
\code{TRUE}, \code{standardize} is ignored. Default is \code{FALSE}.}

\item{standardize}{optional boolean indicating whether to return results for
standardized data. If \code{intercept} is \code{TRUE}, this argument is
ignored. Default is \code{FALSE}.
\code{TRUE}, \code{standardize} is ignored. Default is \code{FALSE}.}

\item{kernel}{optional kernel to use in the ridge regression model. By
default the linear kernel with constant zero is used, that is, no kernel
transformation is applied. See the details section for more details on
available kernel transformations.}

\item{const}{optional constant parameter for the kernel transformation.
Default is \code{0}.}

\item{degree}{optional degree parameter in the kernel transformation. Default
is \code{NULL}.}

\item{scale}{optional scale parameter in the kernel transformation. Default
is \code{NULL}.}
}
\value{
\code{dual.ridge.lm} returns an object of \code{\link{class}}
\code{mlkit.dual.ridge.fit}. An object of class \code{mlkit.dual.ridge.fit}
is a list containing at least the following components:
\item{coefficients}{a named vector of optimal coefficients.}
\item{alpha}{L1-weight hyperparameter in elastic net penalty term.}
\item{lambda}{penalty term scaling hyperparameter.}
\item{r2}{coefficient of determination for optimal coefficients.}
\item{x}{matrix containing the explanatory variables used in estimation.}
\item{ker.mat}{kernel matrix used in estimation.}
\item{kernel}{kernel transformation.}
\item{const}{constant parameter in the kernel transformation.}
\item{degree}{degree parameter in the kernel transformation.}
\item{scale}{scale parameter in the kernel transformation.}
}
\description{
Implementation of the dual analytical solution for a linear
regression model with a Ridge penalty term.
}
