% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic.net.lm.R
\name{elastic.net.lm}
\alias{elastic.net.lm}
\title{Fitting Linear Models with Elastic Net Penalty}
\usage{
elastic.net.lm(
  formula,
  data,
  lambda,
  alpha,
  intercept = F,
  standardize = F,
  beta.init = NULL,
  beta.tol = 0,
  loss.tol = 0.000001,
  eps = 0.000001,
  seed = NULL,
  verbose = 0
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}: a symbolic
description of the model to be fitted following the standard of
\code{\link[stats]{lm}}.}

\item{data}{an optional data frame, list or environment (or object coercible
by \code{\link[base]{as.data.frame}} to a data frame) containing the
variables in the model. If not found in data, the variables are taken from
environment (\code{formula}), typically the environment from which this
function is called.}

\item{lambda}{penalty term scaling hyperparameter.}

\item{alpha}{L1-weight hyperparameter in elastic net penalty term.}

\item{intercept}{optional boolean indicating whether to fit an intercept. If
\code{TRUE}, \code{standardize} is ignored. Default is \code{FALSE}.}

\item{standardize}{optional boolean indicating whether to return results for
standardized data. If \code{intercept} is \code{TRUE}, this argument is
ignored. Default is \code{FALSE}.
\code{TRUE}, \code{standardize} is ignored. Default is \code{FALSE}.}

\item{beta.init}{optional initial beta parameters to use in the MM
algorithm. Default is \code{NULL}.}

\item{beta.tol}{optional absolute tolerance for rounding down parameter
standardized estimates. If the absolute value of a parameter estimate in the
standardized model is smaller than \code{beta.tol}, it is rounded down to
zero. Default is \code{0}, that is, no rounding.}

\item{loss.tol}{optional convergence tolerance on the elastic net loss in
the MM algorithm. Default is \code{1e-6}.}

\item{eps}{optional correction term to avoid rounding by zero. Default is
\code{1e-6}.}

\item{seed}{optional seed. Default is \code{NULL}.}

\item{verbose}{optional number indicating per how many iterations the
estimation progress is displayed. Default is \code{0}, that is, no progress
updates.}
}
\value{
\code{elastic.net.lm} returns an object of \code{\link{class}}
\code{mlkit.lm.fit}. An object of class \code{mlkit.lm.fit} is a list
containing at least the following components:
\item{coefficients}{a named vector of optimal coefficients.}
\item{loss}{residual sum of squares plus elastic net loss for optimal
coefficients.}
\item{r2}{coefficient of determination for optimal coefficients.}
\item{adj.r2}{adjusted coefficient of determination for optimal
coefficients.}
}
\description{
Implementation of the MM algorithm solver for a linear
regression model with an elastic net penalty term.
}
