% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.search.cross.validation.R
\name{grid.search.cross.validation}
\alias{grid.search.cross.validation}
\title{Grid search K-fold cross-validation}
\usage{
grid.search.cross.validation(
  formula,
  data,
  estimator,
  params.list,
  n.folds = 5,
  ind.metric,
  comb.metric = mean,
  fold.id = NULL,
  force = F,
  verbose = F,
  plot = F,
  contour.scale = NULL,
  coef.lims = NULL,
  coef.names = NULL,
  seed = NULL,
  use.formula = T,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}: a symbolic
description of the model to be fitted following the standard of
\code{\link[stats]{lm}}.}

\item{data}{an optional data frame, list or environment (or object coercible
by \code{\link[base]{as.data.frame}} to a data frame) containing the
variables in the model. If not found in data, the variables are taken from
environment (\code{formula}), typically the environment from which this
function is called.}

\item{estimator}{estimator function that has arguments \code{formula} and
\code{data} and returns a list containing parameters estimates in the
component \code{coefficients}.}

\item{params.list}{list or vector containing hyperparameters and their
respective values to consider}

\item{n.folds}{optional number of folds (K) in cross-validation. Default is
5.}

\item{ind.metric}{metric function taking in a numerical vector of
predictions for the dependent variable together with a numerical vector of
actual outcomes of the dependent variable that returns a performance metric
for the individual folds.}

\item{comb.metric}{optional function used to combine individual fold
metrics. Default is \code{\link[base]{mean}}.}

\item{fold.id}{optional vector containing numerical fold identifiers for
each row in the data. If \code{NULL}, \code{n.folds} is used and random fold
identifiers are constructed divided the observations equally over K folds.
If provided, \code{n.folds} is ignored. Default is \code{NULL}.}

\item{force}{optional boolean indicating whether or not to allow for errors
due to singularity when applying the estimator. If \code{TRUE}, the
individual metric of folds with hyperparameter combinations for which the
estimator is not able to estimate coefficients due to singularity, are set to
\code{Inf} and hence the errors are ignored. Default is \code{FALSE}.}

\item{verbose}{optional boolean indicating whether to show a progress bar.
Default is \code{FALSE}.}

\item{plot}{optional boolean indicating whether to generate heatmaps of
performance on the grid and displaying the estimated coefficients. Default
is \code{FALSE}.}

\item{contour.scale}{optional vector containing the hyperparameters and
their scales for the axes in the contour plots. Default is \code{NULL}.}

\item{coef.lims}{optimal limits of the coefficients plot. Default is
\code{NULL}.}

\item{coef.names}{optional names of coefficients, used in estimates barplot.
Default is \code{NULL} in which case the column names of the explanatory
variables are used.}

\item{seed}{optimal seed to specify. Default is \code{NULL}.}

\item{use.formula}{whether or not to use the formula data combination of X
and y as inputs to the model. Default is \code{TRUE}.}

\item{...}{additional arguments to be passed to the \code{estimator}
function.}
}
\value{
\code{grid.search.cross.validation} returns an object of
\code{\link{class}} \code{"gscv"}. An object of class \code{"gscv"} is a
list containing at least the following components:
\item{coefficients}{a named vector of optimal coefficients.}
\item{metric}{metric of the optimal hyperparameters.}
\item{params}{a named vector of optimal hyperparameters.}
}
\description{
Implementation of the grid search approach using K-fold cross-validation for
hyperparameter tuning of a given \code{estimator}.
}
