% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mlkit.class.fit.R
\name{predict.mlkit.class.fit}
\alias{predict.mlkit.class.fit}
\title{Model Predictions for Classifier Models}
\usage{
\method{predict}{mlkit.class.fit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{\code{mlkit.class.fit} object generated by a call to a
classifier method in the \code{mlkit} package that is used for prediction.}

\item{newdata}{optional matrix of explanatory variables to use in prediction.
Default is \code{NULL} in which case the in-sample predictions are returned.}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
Atomic vector containing predictions based on the given model and
explanatory variables.
}
\description{
Custom implementation for prediction using model fits obtained
resulting from multiclass classification models in the \code{mlkit} package.
}
