% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.dual.ridge.lm.R
\name{predict.mlkit.dual.ridge.fit}
\alias{predict.mlkit.dual.ridge.fit}
\title{Model Predictions for Dual Ridge Regression Model}
\usage{
\method{predict}{mlkit.dual.ridge.fit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{\code{mlkit.dual.ridge.fit} object generated by a call to the
\code{\link{dual.ridge.lm}} method that is used for prediction.}

\item{newdata}{optional matrix of explanatory variables to use in prediction.
Default is \code{NULL} in which case the in-sample predictions are returned.}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
Atomic vector containing predictions based on the given model and
explanatory variables.
}
\description{
Custom implementation for prediction using model fits using the
\code{\link{dual.ridge.lm}} method.
}
