% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mlkit.lm.fit.R
\name{predict.mlkit.lm.fit}
\alias{predict.mlkit.lm.fit}
\title{Model Predictions for Linear Regression Model}
\usage{
\method{predict}{mlkit.lm.fit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{\code{mlkit.lm.fit} object generated by a call to a linear
regression method in the \code{mlkit} package that is used for prediction.}

\item{newdata}{optional matrix of explanatory variables to use in prediction.
Default is \code{NULL} in which case the in-sample predictions are returned.}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
Atomic vector containing predictions based on the given model and
explanatory variables.
}
\description{
Custom implementation for prediction using model fits obtained
resulting from linear regression models in the \code{mlkit} package.
}
