% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.bin.R
\name{svm.bin}
\alias{svm.bin}
\title{Fitting Binary Support Vector Machine}
\usage{
svm.bin(
  formula,
  data,
  lambda,
  loss = "qua",
  huber.k = NULL,
  loss.tol = 0.000001,
  v.init = NULL,
  seed = NULL,
  verbose = 0
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}: a symbolic
description of the model to be fitted following the standard of
\code{\link[stats]{lm}}.}

\item{data}{an optional data frame, list or environment (or object coercible
by \code{\link[base]{as.data.frame}} to a data frame) containing the
variables in the model. If not found in data, the variables are taken from
environment (\code{formula}), typically the environment from which this
function is called.}

\item{lambda}{penalty term scaling hyperparameter.}

\item{loss}{hinge loss to use in estimation. Default is \code{'qua'}, that
is, quadratic hinge loss.}

\item{huber.k}{hyperparameter for Huber hinge errors. Default is
\code{NULL}.}

\item{loss.tol}{optional convergence tolerance in the MM algorithm. Default
is \code{1e-6}.}

\item{v.init}{optional initial v parameters to use in the MM algorithm.
Default is \code{NULL}.}

\item{seed}{optional seed. Default is \code{NULL}.}

\item{verbose}{optional number indicating per how many iterations the
estimation progress is displayed. Default is \code{0}, that is, no progress
updates.}
}
\value{
\code{svm.bin} returns an object of \code{\link{class}}
\code{mlkit.bin.fit}. An object of class \code{mlkit.bin.fit} is a list
containing at least the following components:
\item{coefficients}{a named vector of optimal coefficients.}
\item{loss}{SVM loss including penalizing term on the parameters.}
}
\description{
Implementation of the MM algorithm solver for the binary
support vector machine.
}
